Develop

+++ src/tasks/tasks.service.ts
@@ -23,12 +23,16 @@ export class TasksService {
         return this.taskRepository.getTasks(filterDto, user);
     }
     
-    async getTaskByID(@Param('id', ParseIntPipe) id:number): Promise<Task> {
-            const found = await this.taskRepository.findOne(id);
+    async getTaskByID(
+        id:number,
+        user: User,
+    ): Promise<Task> {
+            const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
             if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
             }
+            console.log(found)
             return found;
         }
 
@@ -39,17 +43,17 @@ export class TasksService {
         return this.taskRepository.createTask(createTaskDto, user);
     } 
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
+        console.log(result)
         // console.log(result);
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
-
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus) {
-        const task = await this.getTaskByID(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User) {
+        const task = await this.getTaskByID(id, user);
         task.status = status;
         await task.save();
         return task;
